Extract option utils into a new file
